[frontend] Refactor profile page
