+++ frontend/app/profile/layout.tsx
@@ -1,5 +1,8 @@
+"use client";
+import { SidebarNav } from "@/app/ui/profile/sidebar-nav";
+import { Stack } from "@/app/ui/layout/stack";
 import { Separator } from "@/components/ui/separator";
-import { SidebarNav } from "@/app/ui/sidebar-nav";
+import { usePathname } from "next/navigation";
 
 const sidebarNavItems = [
   {
@@ -33,24 +36,24 @@ interface SettingsLayoutProps {
 }
 
 export default function SettingsLayout({ children }: SettingsLayoutProps) {
+  const pathname = usePathname();
+  const title = sidebarNavItems.find((item) => item.href === pathname)?.title;
   return (
     <>
-      <div className="md:hidden">
-        <div className="space-y-6 p-10 pb-16">Can&apos;t be displayed.</div>
-      </div>
-      <div className="hidden pb-16 md:block">
-        <div className="space-y-0.5">
-          <h2 className="text-2xl font-bold tracking-tight">Settings</h2>
-          <p className="text-muted-foreground">
-            Manage your account settings and set e-mail preferences.
-          </p>
-        </div>
-        <Separator className="my-6" />
+      <div className="pb-16">
         <div className="flex flex-col space-y-8 lg:flex-row lg:space-x-12 lg:space-y-0">
           <aside className="lg:w-1/5">
             <SidebarNav items={sidebarNavItems} />
           </aside>
-          <div className="flex-1 lg:max-w-2xl">{children}</div>
+          <div className="flex-1 lg:max-w-2xl">
+            <Stack spacing={4}>
+              <Stack spacing={1}>
+                <div className="text-2xl">{title}</div>
+                <Separator />
+              </Stack>
+              {children}
+            </Stack>
+          </div>
         </div>
       </div>
     </>
